add exchangeCurrency function, currency code table and start rearrange function in different modules
