+++ cuntryCurrency.json
@@ -1,0 +1,271 @@
+{
+    "EMBER COUNTRIES OF THE AFRICAN DEVELOPMENT BANK GROUP":"XUA",
+    "AFGHANISTAN":"AFN",
+    "ALGERIA":"DZD",
+    "ARGENTINA":"ARS",
+    "ARMENIA":"AMD",
+    "ARUBA":"AWG",
+    "AUSTRALIA":"AUD",
+    "CHRISTMAS ISLAND":"AUD",
+    "COCOS (KEELING) ISLANDS (THE)":"AUD",
+    "HEARD ISLAND AND McDONALD ISLANDS":"AUD",
+    "KIRIBATI":"AUD",
+    "NAURU":"AUD",
+    "NORFOLK ISLAND":"AUD",
+    "TUVALU":"AUD",
+    "AZERBAIJAN":"AZN",
+    "BAHAMAS (THE)":"BSD",
+    "BAHRAIN":"BHD",
+    "THAILAND":"THB",
+    "PANAM":"PAB",
+    "BARBADOS":"BBD",
+    "BELARUS":"BYN",
+    "BELIZE":"BZD",
+    "BERMUDA":"BMD",
+    "VENEZUELA (BOLIVARIAN REPUBLIC OF)":"VEF",
+    "VENEZUELA (BOLIVARIAN REPUBLIC O)":"VED",
+    "BOLIVIA (PLURINATIONAL STATE O)":"BOB",
+    "BRAZIL":"BRL",
+    "BRUNEI DARUSSALAM":"BND",
+    "BULGARIA":"BGN",
+    "BURUNDI":"BIF",
+    "CABO VERDE":"CVE",
+    "CANADA":"CAD",
+    "CAYMAN ISLANDS (THE)":"KYD",
+    "BENIN":"XOF",
+    "BURKINA FASO":"XOF",
+    "CÔTE D'IVOIRE":"XOF",
+    "GUINEA-BISSAU":"XOF",
+    "MALI":"XOF",
+    "NIGER (THE)":"XOF",
+    "SENEGAL":"XOF",
+    "TOGO":"XOF",
+    "CAMEROON":"XAF",
+    "CENTRAL AFRICAN REPUBLIC (THE)":"XAF",
+    "CHAD":"XAF",
+    "CONGO (THE)":"XAF",
+    "EQUATORIAL GUINEA":"XAF",
+    "GABON":"XAF",
+    "FRENCH POLYNESIA":"XPF",
+    "NEW CALEDONIA":"XPF",
+    "WALLIS AND FUTUNA":"XPF",
+    "CHIL":"CLP",
+    "COLOMBI":"COP",
+    "COMOROS (THE)":"KMF",
+    "CONGO (THE DEMOCRATIC REPUBLIC OF THE)":"CDF",
+    "BOSNIA AND HERZEGOVINA":"BAM",
+    "NICARAGUA":"NIO",
+    "COSTA RICA":"CRC",
+    "CUB":"CUP",
+    "CZECH REPUBLIC (THE)":"CZK",
+    "GAMBIA (THE)":"GMD",
+    "DENMARK":"DKK",
+    "FAROE ISLANDS (THE)":"DKK",
+    "GREENLAND":"DKK",
+    "REPUBLIC OF NORTH MACEDONIA":"MKD",
+    "DJIBOUTI":"DJF",
+    "SAO TOME AND PRINCIPE":"STN",
+    "DOMINICAN REPUBLIC (THE)":"DOP",
+    "VIET NAM":"VND",
+    "ANGUILLA":"XCD",
+    "ANTIGUA AND BARBUDA":"XCD",
+    "DOMINICA":"XCD",
+    "GRENADA":"XCD",
+    "MONTSERRAT":"XCD",
+    "SAINT KITTS AND NEVIS":"XCD",
+    "SAINT LUCIA":"XCD",
+    "SAINT VINCENT AND THE GRENADINES":"XCD",
+    "EGYPT":"EGP",
+    "EL SALVADO":"SVC",
+    "ETHIOPIA":"ETB",
+    "ÅLAND ISLANDS":"EUR",
+    "ANDORRA":"EUR",
+    "AUSTRIA":"EUR",
+    "BELGIUM":"EUR",
+    "CYPRUS":"EUR",
+    "ESTONIA":"EUR",
+    "EUROPEAN UNION":"EUR",
+    "FINLAND":"EUR",
+    "FRANCE":"EUR",
+    "FRENCH GUIANA":"EUR",
+    "FRENCH SOUTHERN TERRITORIES (THE)":"EUR",
+    "GERMANY":"EUR",
+    "GREECE":"EUR",
+    "GUADELOUPE":"EUR",
+    "HOLY SEE (THE)":"EUR",
+    "IRELAND":"EUR",
+    "ITALY":"EUR",
+    "LATVIA":"EUR",
+    "LITHUANIA":"EUR",
+    "LUXEMBOURG":"EUR",
+    "MALTA":"EUR",
+    "MARTINIQUE":"EUR",
+    "MAYOTTE":"EUR",
+    "MONACO":"EUR",
+    "MONTENEGRO":"EUR",
+    "NETHERLANDS (THE)":"EUR",
+    "PORTUGAL":"EUR",
+    "RÉUNION":"EUR",
+    "SAINT BARTHÉLEMY":"EUR",
+    "SAINT MARTIN (FRENCH PART)":"EUR",
+    "SAINT PIERRE AND MIQUELON":"EUR",
+    "SAN MARINO":"EUR",
+    "SLOVAKIA":"EUR",
+    "SLOVENIA":"EUR",
+    "SPAIN":"EUR",
+    "FALKLAND ISLANDS (THE) [MALVINAS]":"FKP",
+    "FIJI":"FJD",
+    "HUNGARY":"HUF",
+    "GHANA":"GHS",
+    "GIBRALTAR":"GIP",
+    "HAIT":"HTG",
+    "PARAGUAY":"PYG",
+    "GUINEA":"GNF",
+    "GUYANA":"GYD",
+    "HONG KONG":"HKD",
+    "UKRAINE":"UAH",
+    "ICELAND":"ISK",
+    "BHUTA":"INR",
+    "INDIA":"INR",
+    "IRAN (ISLAMIC REPUBLIC OF)":"IRR",
+    "IRAQ":"IQD",
+    "JAMAICA":"JMD",
+    "JORDAN":"JOD",
+    "KENYA":"KES",
+    "PAPUA NEW GUINEA":"PGK",
+    "LAO PEOPLE’S DEMOCRATIC REPUBLIC (THE)":"LAK",
+    "CROATIA":"HRK",
+    "KUWAIT":"KWD",
+    "MALAWI":"MWK",
+    "ANGOLA":"AOA",
+    "MYANMAR":"MMK",
+    "GEORGIA":"GEL",
+    "LEBANON":"LBP",
+    "ALBANIA":"ALL",
+    "HONDURAS":"HNL",
+    "SIERRA LEONE":"SLE",
+    "LIBERIA":"LRD",
+    "LIBYA":"LYD",
+    "SWAZILAND":"SZL",
+    "LESOTH":"LSL",
+    "MADAGASCAR":"MGA",
+    "MALAYSIA":"MYR",
+    "MAURITIUS":"MUR",
+    "MEXIC":"MXN",
+    "MEXICO":"MXV",
+    "MOLDOVA (THE REPUBLIC OF)":"MDL",
+    "MOROCCO":"MAD",
+    "WESTERN SAHARA":"MAD",
+    "MOZAMBIQUE":"MZN",
+    "BOLIVIA (PLURINATIONAL STATE OF)":"BOV",
+    "NIGERIA":"NGN",
+    "ERITREA":"ERN",
+    "NAMIBI":"NAD",
+    "NEPAL":"NPR",
+    "CURAÇAO":"ANG",
+    "SINT MAARTEN (DUTCH PART)":"ANG",
+    "ISRAEL":"ILS",
+    "TAIWAN (PROVINCE OF CHINA)":"TWD",
+    "COOK ISLANDS (THE)":"NZD",
+    "NEW ZEALAND":"NZD",
+    "NIUE":"NZD",
+    "PITCAIRN":"NZD",
+    "TOKELAU":"NZD",
+    "BHUTAN":"BTN",
+    "ANTARCTICA":"",
+    "PALESTINE, STATE OF":"",
+    "SOUTH GEORGIA AND THE SOUTH SANDWICH ISLANDS":"",
+    "KOREA (THE DEMOCRATIC PEOPLE’S REPUBLIC OF)":"KPW",
+    "BOUVET ISLAND":"NOK",
+    "NORWAY":"NOK",
+    "SVALBARD AND JAN MAYEN":"NOK",
+    "PERU":"PEN",
+    "MAURITANIA":"MRU",
+    "PAKISTAN":"PKR",
+    "MACAO":"MOP",
+    "TONGA":"TOP",
+    "CUBA":"CUC",
+    "URUGUA":"UYU",
+    "PHILIPPINES (THE)":"PHP",
+    "GUERNSEY":"GBP",
+    "ISLE OF MAN":"GBP",
+    "JERSEY":"GBP",
+    "UNITED KINGDOM OF GREAT BRITAIN AND NORTHERN IRELAND (THE)":"GBP",
+    "BOTSWANA":"BWP",
+    "QATAR":"QAR",
+    "GUATEMALA":"GTQ",
+    "LESOTHO":"ZAR",
+    "NAMIBIA":"ZAR",
+    "SOUTH AFRICA":"ZAR",
+    "OMAN":"OMR",
+    "CAMBODIA":"KHR",
+    "ROMANIA":"RON",
+    "MALDIVES":"MVR",
+    "INDONESIA":"IDR",
+    "RUSSIAN FEDERATION (THE)":"RUB",
+    "RWANDA":"RWF",
+    "SAINT HELENA, ASCENSION AND TRISTAN DA CUNHA":"SHP",
+    "SAUDI ARABIA":"SAR",
+    "INTERNATIONAL MONETARY FUND (IMF) ":"XDR",
+    "SERBIA":"RSD",
+    "SEYCHELLES":"SCR",
+    "SINGAPORE":"SGD",
+    "SOLOMON ISLANDS":"SBD",
+    "KYRGYZSTAN":"KGS",
+    "SOMALIA":"SOS",
+    "TAJIKISTAN":"TJS",
+    "SOUTH SUDAN":"SSP",
+    "SRI LANKA":"LKR",
+    "SISTEMA UNITARIO DE COMPENSACION REGIONAL DE PAGOS 'SUCRE'":"XSU",
+    "SUDAN (THE)":"SDG",
+    "SURINAME":"SRD",
+    "SWEDEN":"SEK",
+    "LIECHTENSTEIN":"CHF",
+    "SWITZERLAN":"CHF",
+    "SYRIAN ARAB REPUBLIC":"SYP",
+    "BANGLADESH":"BDT",
+    "SAMOA":"WST",
+    "TANZANIA, UNITED REPUBLIC OF":"TZS",
+    "KAZAKHSTAN":"KZT",
+    "TRINIDAD AND TOBAGO":"TTD",
+    "MONGOLIA":"MNT",
+    "TUNISIA":"TND",
+    "TURKEY":"TRY",
+    "TURKMENISTAN":"TMT",
+    "UNITED ARAB EMIRATES (THE)":"AED",
+    "UGANDA":"UGX",
+    "CHILE":"CLF",
+    "COLOMBIA":"COU",
+    "URUGUAY":"UYI",
+    "AMERICAN SAMOA":"USD",
+    "BONAIRE, SINT EUSTATIUS AND SABA":"USD",
+    "BRITISH INDIAN OCEAN TERRITORY (THE)":"USD",
+    "ECUADOR":"USD",
+    "EL SALVADOR":"USD",
+    "GUAM":"USD",
+    "HAITI":"USD",
+    "MARSHALL ISLANDS (THE)":"USD",
+    "MICRONESIA (FEDERATED STATES OF)":"USD",
+    "NORTHERN MARIANA ISLANDS (THE)":"USD",
+    "PALAU":"USD",
+    "PANAMA":"USD",
+    "PUERTO RICO":"USD",
+    "TIMOR-LESTE":"USD",
+    "TURKS AND CAICOS ISLANDS (THE)":"USD",
+    "UNITED STATES MINOR OUTLYING ISLANDS (THE)":"USD",
+    "UNITED STATES OF AMERICA (THE":"USD",
+    "VIRGIN ISLANDS (BRITISH)":"USD",
+    "VIRGIN ISLANDS (U.S.)":"USD",
+    "UNITED STATES OF AMERICA (THE)":"USN",
+    "UZBEKISTAN":"UZS",
+    "VANUATU":"VUV",
+    "SWITZERLAND":"CHE",
+    "SWITZERLA":"CHW",
+    "KOREA (THE REPUBLIC OF)":"KRW",
+    "YEMEN":"YER",
+    "JAPAN":"JPY",
+    "CHINA":"CNY",
+    "ZAMBIA":"ZMW",
+    "ZIMBABWE":"ZWL",
+    "POLAND":"PLN"    
+}
